+++ templates/data/options.py
@@ -13,9 +13,13 @@ Settings and options for the bot.
 ---
 """
 
+
 # The prefix for the bot (str). Define a list of stings for multiple prefixes.
 # ie: `["?", "!", "pls "]`
-command_prefix = "mb?"
+command_prefix = ["mb? ", "mb?", "Mb? ", "Mb?"]
+
+# The prefix to use for display purposes (ex: status message).
+display_prefix = "mb?"
 
 
 # The roles to give to new members, per guild.