+++ info.py
@@ -7,5 +7,5 @@ author = "@0x5c#0639"
 description = """A bot for basic guild stuff."""
 license = "MIT"
 contributing = "https://github.com/0x5c/minibot"
-release = '1.2.1'
+release = '1.3.0'
 guild_invite = "https://discord.gg/Wq8vpm3"